requirements.txt
streamlit
langchain
chromadb
FlagEmbedding
python-dotenv
redis
openai